+++ analysis/src/BuildSystem.ml
@@ -5,23 +5,24 @@ let namespacedName namespace name =
 
 let ( /+ ) = Filename.concat
 
-let getBsPlatformDir rootPath =
+let getRuntimeDir rootPath =
   match !Cfg.isDocGenFromCompiler with
   | false -> (
     let result =
-      ModuleResolution.resolveNodeModulePath ~startPath:rootPath "rescript"
+      ModuleResolution.resolveNodeModulePath ~startPath:rootPath
+        "@rescript/runtime"
     in
     match result with
     | Some path -> Some path
     | None ->
-      let message = "rescript could not be found" in
+      let message = "@rescript/runtime could not be found" in
       Log.log message;
       None)
   | true -> Some rootPath
 
-let getLibBs root = Files.ifExists (root /+ "lib" /+ "bs")
+let getLibBs path = Files.ifExists (path /+ "lib" /+ "bs")
 
 let getStdlib base =
-  match getBsPlatformDir base with
+  match getRuntimeDir base with
   | None -> None
-  | Some bsPlatformDir -> Some (bsPlatformDir /+ "lib" /+ "ocaml")
+  | Some runtimeDir -> Some (runtimeDir /+ "lib" /+ "ocaml")