+++ cli/common/bsb.js
@@ -5,9 +5,9 @@ import * as fs from "node:fs";
 import { createServer } from "node:http";
 import * as os from "node:os";
 import * as path from "node:path";
-import { WebSocket } from "#lib/minisocket";
 
 import { rescript_legacy_exe } from "./bins.js";
+import { WebSocket } from "./minisocket.js";
 
 const cwd = process.cwd();
 const lockFileName = path.join(cwd, ".bsb.lock");