+++ compiler/bsc/rescript_compiler_main.ml
@@ -205,12 +205,6 @@ let eval (s : string) ~suffix =
 
 (* let (//) = Filename.concat *)
 
-let print_standard_library () =
-  let standard_library = Config.standard_library in
-  print_string standard_library;
-  print_newline ();
-  exit 0
-
 let bs_version_string = "ReScript " ^ Bs_version.version
 
 let print_version_string () =
@@ -378,9 +372,6 @@ let command_line_flags : (string * Bsc_args.spec * string) array =
     ( "-ignore-parse-errors",
       set Clflags.ignore_parse_errors,
       "*internal* continue after parse errors" );
-    ( "-where",
-      unit_call print_standard_library,
-      "*internal* Print location of standard library and exit" );
     ( "-verbose",
       set Clflags.verbose,
       "*internal* Print calls to external commands" );