+++ compiler/ext/config.mli
@@ -15,11 +15,14 @@
 
 (* System configuration *)
 
-val standard_library : string
-(* The directory containing the standard libraries *)
+(* The directory containing the runtime module (@rescript/runtime) *)
+val runtime_module_path : string
 
-val load_path : string list ref
+(* The directory containing the runtime artifacts (@rescript/runtime/lib/ocaml) *)
+val standard_library : string
+
 (* Directories in the search path for .cmi and .cmo files *)
+val load_path : string list ref
 
 val cmi_magic_number : string