+++ tests/build_tests/jsx_settings_inheritance/input.js
@@ -4,5 +4,5 @@ import { setup } from "#dev/process";
 
 const { execBuild, execClean } = setup(import.meta.dirname);
 
-await execClean()
+await execClean();
 await execBuild();