+++ tests/build_tests/super_errors/input.js
@@ -11,7 +11,7 @@ const { bsc } = setup(import.meta.dirname);
 const expectedDir = path.join(import.meta.dirname, "expected");
 
 const fixtures = readdirSync(path.join(import.meta.dirname, "fixtures")).filter(
-  (fileName) => path.extname(fileName) === ".res"
+  fileName => path.extname(fileName) === ".res",
 );
 
 const prefix = ["-w", "+A", "-bs-jsx", "4"];